modify node build

+++ templates/jquery-handlebars.js
@@ -16,39 +16,45 @@ module.exports = async({ name, version, author }) => {
     default: false
   }])
   let spinner = ora(`Initializing the ${name} project\n`).start()
-  let packageJSON = Handlebars.compile(JSON.stringify(deepmerge(require('./template.json'), deepmerge({
-    'scripts': {
-      'server': 'manpacker server --ic ./config/localserver.ic.js',
-      'dev': 'manpacker build ./config/development.ic.js',
-      'build': 'manpacker build ./config/production.ic.js'
-    },
+  let packageJSON = Handlebars.compile(JSON.stringify(
+    deepmerge(require('./template.json'), deepmerge({
+      'scripts': {
+        'start': 'npm run server',
+        'server': 'manpacker server --ic ./config/localserver.ic.js',
+        'dev': 'manpacker build ./config/development.ic.js',
+        'build': 'manpacker build ./config/production.ic.js'
+      },
 
-    'devDependencies': {
-      '@manpacker/generator': await packer.version(
-        `${npms}/${encodeURIComponent('@manpacker/generator')}`
-      ),
-      'babel-preset-manpacker': await packer.version(
-        `${npms}/babel-preset-manpacker`
-      ),
-      'eslint': '^5.16.0',
-      'eslint-config-manpacker': await packer.version(
-        `${npms}/eslint-config-manpacker`
-      )
-    },
+      'devDependencies': {
+        '@manpacker/generator': await packer.version(
+          `${npms}/${encodeURIComponent('@manpacker/generator')}`
+        ),
+        'babel-preset-manpacker': await packer.version(
+          `${npms}/babel-preset-manpacker`
+        ),
+        'eslint': '^5.16.0',
+        'eslint-config-manpacker': await packer.version(
+          `${npms}/eslint-config-manpacker`
+        )
+      },
 
-    'babel': { 'presets': ['manpacker'] },
+      'babel': { 'presets': ['manpacker'] },
 
-    'eslintConfig': { 'extends': ['manpacker'] },
+      'eslintConfig': {
+        'extends': (isNode ? ['manpacker-typescript'] : []).concat(['manpacker'])
+      },
 
-    dependencies: {
-      jquery: await packer.version(`${npms}/jquery`),
-      handlebars: await packer.version(`${npms}/handlebars`)
-    }
-  }, isNode ? await require('./noden')() : {})), null, '  '))({ name, version, author })
+      dependencies: {
+        jquery: await packer.version(`${npms}/jquery`),
+        handlebars: await packer.version(`${npms}/handlebars`)
+      }
+    }, isNode ? await require('./noden')() : {})), null, '  '))({ name, version, author })
 
   packer.clone(`${org}/jquery-handlebars.git`, name)
-  isNode && rimraf.sync(path.resolve(`${name}/app`)) &&
-  rimraf.sync(path.resolve(`${name}/tsconfig.json`))
+  if (!isNode) {
+    rimraf.sync(path.resolve(`${name}/app`))
+    rimraf.sync(path.resolve(`${name}/tsconfig.json`))
+  }
   packer.write(`${name}/package.json`, packageJSON)
   spinner.stop()
   packer.install(name)
